add DAO scripts
