+++ test/integration/proposals_config.json
@@ -1,7 +1,7 @@
 {
-    "fip_28" : {
-        "deploy" : false,
-        "exec" : true,
+    "governorUpgrade" : {
+        "deploy" : true,
+        "exec" : false,
         "proposerAddress" : "0xb81cf4981Ef648aaA73F07a18B03970f04d5D8bF"
     },
     "v2Phase1": {
